finish permission gate

+++ app/Http/Controllers/BookingController.php
@@ -16,6 +16,7 @@ class BookingController extends Controller
 {
     public function index(): Factory|View|Application
     {
+        abort_unless(auth()->user()->can('View Booking'), 403);
         $bookings = Booking::with(['user', 'package', 'package.tour'])->orderByDesc('id')->paginate(10);
         $setting = app(GeneralSetting::class);
 
@@ -24,11 +25,13 @@ class BookingController extends Controller
 
     public function create(): Factory|View|Application
     {
+        abort_unless(auth()->user()->can('Create Booking'), 403);
         return view('smartTT.booking.create');
     }
 
     public function show(Booking $booking): Factory|View|Application
     {
+        abort_unless(auth()->user()->can('View Booking'), 403);
         $booking->load(['user', 'package', 'package.tour', 'payment', 'guests']);
         $setting = app(GeneralSetting::class);
         $bookingSetting = app(BookingSetting::class);
@@ -38,6 +41,7 @@ class BookingController extends Controller
 
     public function destroy(Booking $booking): RedirectResponse
     {
+        abort_unless(auth()->user()->can('Delete Booking'), 403);
         $booking->delete();
 
         return redirect()->route('bookings.index')->with('success', __('Booking deleted successfully'));
@@ -45,7 +49,8 @@ class BookingController extends Controller
 
     public function addPayment(Booking $booking)
     {
-        $paymentAmount = $booking->total_price - $booking->payment->filter(fn (Payment $payment) => $payment->status === Payment::STATUS_PAID)->sum('amount');
+        abort_unless(auth()->user()->can('Create Payment'), 403);
+        $paymentAmount = $booking->total_price - $booking->payment->filter(fn(Payment $payment) => $payment->status === Payment::STATUS_PAID)->sum('amount');
         if ($paymentAmount <= 0) {
             return redirect()->route('bookings.show', $booking)->withErrors(__('Payment already completed'));
         }
@@ -55,6 +60,7 @@ class BookingController extends Controller
 
     public function audit(Booking $booking)
     {
+        abort_unless(auth()->user()->can('Audit Booking'), 403);
         $logs = Activity::forSubject($booking)->get();
 
         return view('smartTT.booking.audit', compact('logs', 'booking'));
