+++ app/Http/Controllers/TourDescriptionController.php
@@ -17,6 +17,7 @@ class TourDescriptionController extends Controller
 {
     public function update(Request $request, TourDescription $tourDescription, UpdateTourDescriptionAction $action): RedirectResponse
     {
+        abort_unless(auth()->user()->can('Edit Tour Description'), 403);
         $action->execute($request->all(), $tourDescription);
 
         return redirect()->route('tours.show', $tourDescription->tour)
@@ -25,11 +26,14 @@ class TourDescriptionController extends Controller
 
     public function edit(TourDescription $tourDescription): Factory|View|Application
     {
+        abort_unless(auth()->user()->can('Edit Tour Description'), 403);
+
         return view('smartTT.tourDescription.edit', compact('tourDescription'));
     }
 
     public function destroy(TourDescription $tourDescription): RedirectResponse
     {
+        abort_unless(auth()->user()->can('Delete Tour Description'), 403);
         $tourDescription->delete();
 
         return back()->with('success', __('Tour description deleted successfully'));
@@ -37,6 +41,7 @@ class TourDescriptionController extends Controller
 
     public function attachToTour(Request $request, Tour $tour, AttachDescriptionToTourAction $action): RedirectResponse
     {
+        abort_unless(auth()->user()->can('Edit Tour'), 403);
         $action->execute($request->all(), $tour);
 
         return back()->with('success', __('Tour description attached successfully'));
@@ -44,6 +49,7 @@ class TourDescriptionController extends Controller
 
     public function audit(TourDescription $tourDescription)
     {
+        abort_unless(auth()->user()->can('Audit Tour Description'), 403);
         $logs = Activity::forSubject($tourDescription)->get();
 
         return view('smartTT.tourDescription.audit', compact('logs', 'tourDescription'));
